pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aichatbot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "python-telegram-bot>=21.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "pyowm>=3.3.0",
]

[project.scripts]
aichatbot = "aichatbot.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/aichatbot"]
